There\'s an innumerable number of changes that have been made since the last push, for one, and for another, this repo is still in development/immature/etc. Notably, there is an unaccounted for difference between the expected absorption from a neutron\'s path through a material compared to the output result of the step-wise method. I am still searching for it - it\'s not from the (lack of) randomness in the velocities used for simulation, so I have no clue where it really could be from. The effect fundamentally amounts to more abosrption than expected when viewing 1 - n - n\', and more neutrons missing when viewing 1 - n (which is all that is actually observable). All I can think of at the moment is that, somehow, there's something more afoot?

+++ src/lib/material_list.f90
@@ -3,9 +3,9 @@ module material_list
     type materiallist
         character(32)   :: matName, matRange  ! Cols 1, 3
             ! cols          2, 4,          5, 6,   7,   8,    9,      10,  
-        real(8)        ::   d, numDensity, V, Wth, Wsc, Wabs, sigabs, sigel
+        real        ::   d, numDensity, V, Wth, Wsc, Wabs, sigabs, sigel
             ! cols          11       12
-        real(8)        ::   elscatl, abslngth
+        real        ::   elscatl, abslngth
     end type materiallist
 
 contains
@@ -68,7 +68,7 @@ contains
     subroutine get_velocities(INFILE_3, num_vels, vel_list)
         implicit none
         character(64), intent(in)   :: INFILE_3
-        real(8), dimension(:)       :: vel_list
+        real, dimension(:)       :: vel_list
         integer                     :: num_vels, l, error
 
         error = 0
